Incorporate notification from nat_hole_punch crate
