+++ src/handler/request_call.rs
@@ -1,7 +1,7 @@
 pub use crate::node_info::{NodeAddress, NodeContact};
 use crate::{
     packet::Packet,
-    rpc::{Request, RequestBody},
+    rpc::{Payload, Request, RequestBody},
 };
 
 use super::HandlerReqId;
