+++ src/rpc/mod.rs
@@ -1,8 +1,5 @@
-use enr::{CombinedKey, Enr};
 use parse_display_derive::Display;
 use rlp::{DecoderError, Rlp};
-use std::net::{IpAddr, Ipv6Addr};
-use tracing::{debug, warn};
 
 mod notification;
 mod request;
@@ -12,6 +9,7 @@ pub use notification::Notification;
 pub use request::{Request, RequestBody, RequestId};
 pub use response::{Response, ResponseBody};
 
+/// Message type IDs.
 /// Ping notification type.
 pub const PING_MSG_TYPE: u8 = 1;
 /// Pong notification type.
@@ -25,18 +23,25 @@ pub const TALKREQ_MSG_TYPE: u8 = 5;
 /// TalkResp notification type.
 pub const TALKRESP_MSG_TYPE: u8 = 6;
 /// RelayInit notification type.
-pub const REALYINIT_MSG_TYPE: u8 = 7;
+pub const RELAYINIT_MSG_TYPE: u8 = 7;
 /// RelayMsg notification type.
-pub const REALYMSG_MSG_TYPE: u8 = 8;
+pub const RELAYMSG_MSG_TYPE: u8 = 8;
 
-pub trait Payload where Self: Sized {
+/// The payload of message containers SessionMessage, Message or Handshake type.
+pub trait Payload
+where
+    Self: Sized,
+{
+    /// Matches a payload type to its message type id.
     fn msg_type(&self) -> u8;
+    /// Encodes a message to RLP-encoded bytes.
     fn encode(self) -> Vec<u8>;
+    /// Decodes RLP-encoded bytes into a message.
     fn decode(msg_type: u8, rlp: &Rlp<'_>) -> Result<Self, DecoderError>;
 }
 
 #[derive(Debug, Clone, PartialEq, Eq, Display)]
-/// A combined type representing the packet payloads.
+/// A combined type representing the messages which are the payloads of packets.
 pub enum Message {
     /// A request, which contains its [`RequestId`].
     #[display("{0}")]
@@ -48,6 +53,20 @@ pub enum Message {
     #[display("{0}")]
     Notification(Notification),
 }
+
+macro_rules! impl_from_variant_wrap {
+    ($from_type: ty, $to_type: ty, $variant: path) => {
+        impl From<$from_type> for $to_type {
+            fn from(t: $from_type) -> Self {
+                $variant(t)
+            }
+        }
+    };
+}
+
+impl_from_variant_wrap!(Request, Message, Self::Request);
+impl_from_variant_wrap!(Response, Message, Self::Response);
+impl_from_variant_wrap!(Notification, Message, Self::Notification);
 
 #[allow(dead_code)]
 impl Message {
@@ -63,170 +82,21 @@ impl Message {
         if data.len() < 3 {
             return Err(DecoderError::RlpIsTooShort);
         }
-
         let msg_type = data[0];
-
         let rlp = rlp::Rlp::new(&data[1..]);
 
-        let list_len = rlp.item_count().and_then(|size| {
-            if size < 2 {
-                Err(DecoderError::RlpIncorrectListLen)
-            } else {
-                Ok(size)
+        match msg_type {
+            PING_MSG_TYPE | FINDNODE_MSG_TYPE | TALKREQ_MSG_TYPE => {
+                Ok(Request::decode(msg_type, &rlp)?.into())
             }
-        })?;
-
-        let id = RequestId::decode(rlp.val_at::<Vec<u8>>(0)?)?;
-
-        let message = match msg_type {
-            1 => {
-                // PingRequest
-                if list_len != 2 {
-                    debug!(
-                        "Ping Request has an invalid RLP list length. Expected 2, found {}",
-                        list_len
-                    );
-                    return Err(DecoderError::RlpIncorrectListLen);
-                }
-                Payload::Request(Request {
-                    id,
-                    body: RequestBody::Ping {
-                        enr_seq: rlp.val_at::<u64>(1)?,
-                    },
-                })
+            PONG_MSG_TYPE | NODES_MSG_TYPE | TALKRESP_MSG_TYPE => {
+                Ok(Response::decode(msg_type, &rlp)?.into())
             }
-            2 => {
-                // PingResponse
-                if list_len != 4 {
-                    debug!(
-                        "Ping Response has an invalid RLP list length. Expected 4, found {}",
-                        list_len
-                    );
-                    return Err(DecoderError::RlpIncorrectListLen);
-                }
-                let ip_bytes = rlp.val_at::<Vec<u8>>(2)?;
-                let ip = match ip_bytes.len() {
-                    4 => {
-                        let mut ip = [0u8; 4];
-                        ip.copy_from_slice(&ip_bytes);
-                        IpAddr::from(ip)
-                    }
-                    16 => {
-                        let mut ip = [0u8; 16];
-                        ip.copy_from_slice(&ip_bytes);
-                        let ipv6 = Ipv6Addr::from(ip);
-                        // If the ipv6 is ipv4 compatible/mapped, simply return the ipv4.
-                        if let Some(ipv4) = ipv6.to_ipv4() {
-                            IpAddr::V4(ipv4)
-                        } else {
-                            IpAddr::V6(ipv6)
-                        }
-                    }
-                    _ => {
-                        debug!("Ping Response has incorrect byte length for IP");
-                        return Err(DecoderError::RlpIncorrectListLen);
-                    }
-                };
-                let port = rlp.val_at::<u16>(3)?;
-                Payload::Response(Response {
-                    id,
-                    body: ResponseBody::Pong {
-                        enr_seq: rlp.val_at::<u64>(1)?,
-                        ip,
-                        port,
-                    },
-                })
+            RELAYINIT_MSG_TYPE | RELAYMSG_MSG_TYPE => {
+                Ok(Notification::decode(msg_type, &rlp)?.into())
             }
-            3 => {
-                // FindNodeRequest
-                if list_len != 2 {
-                    debug!(
-                        "FindNode Request has an invalid RLP list length. Expected 2, found {}",
-                        list_len
-                    );
-                    return Err(DecoderError::RlpIncorrectListLen);
-                }
-                let distances = rlp.list_at::<u64>(1)?;
-
-                for distance in distances.iter() {
-                    if distance > &256u64 {
-                        warn!(
-                            "Rejected FindNode request asking for unknown distance {}, maximum 256",
-                            distance
-                        );
-                        return Err(DecoderError::Custom("FINDNODE request distance invalid"));
-                    }
-                }
-
-                Payload::Request(Request {
-                    id,
-                    body: RequestBody::FindNode { distances },
-                })
-            }
-            4 => {
-                // NodesResponse
-                if list_len != 3 {
-                    debug!(
-                        "Nodes Response has an invalid RLP list length. Expected 3, found {}",
-                        list_len
-                    );
-                    return Err(DecoderError::RlpIncorrectListLen);
-                }
-
-                let nodes = {
-                    let enr_list_rlp = rlp.at(2)?;
-                    if enr_list_rlp.is_empty() {
-                        // no records
-                        vec![]
-                    } else {
-                        enr_list_rlp.as_list::<Enr<CombinedKey>>()?
-                    }
-                };
-                Payload::Response(Response {
-                    id,
-                    body: ResponseBody::Nodes {
-                        total: rlp.val_at::<u64>(1)?,
-                        nodes,
-                    },
-                })
-            }
-            5 => {
-                // Talk Request
-                if list_len != 3 {
-                    debug!(
-                        "Talk Request has an invalid RLP list length. Expected 3, found {}",
-                        list_len
-                    );
-                    return Err(DecoderError::RlpIncorrectListLen);
-                }
-                let protocol = rlp.val_at::<Vec<u8>>(1)?;
-                let request = rlp.val_at::<Vec<u8>>(2)?;
-                Payload::Request(Request {
-                    id,
-                    body: RequestBody::TalkReq { protocol, request },
-                })
-            }
-            6 => {
-                // Talk Response
-                if list_len != 2 {
-                    debug!(
-                        "Talk Response has an invalid RLP list length. Expected 2, found {}",
-                        list_len
-                    );
-                    return Err(DecoderError::RlpIncorrectListLen);
-                }
-                let response = rlp.val_at::<Vec<u8>>(1)?;
-                Payload::Response(Response {
-                    id,
-                    body: ResponseBody::TalkResp { response },
-                })
-            }
-            _ => {
-                return Err(DecoderError::Custom("Unknown RPC message type"));
-            }
-        };
-
-        Ok(message)
+            _ => Err(DecoderError::Custom("Unknown RPC message type")),
+        }
     }
 }
 
@@ -234,13 +104,15 @@ impl Message {
 mod tests {
     use super::*;
     use enr::EnrBuilder;
+    use enr::{CombinedKey, Enr};
+    use std::net::IpAddr;
 
     #[test]
     fn ref_test_encode_request_ping() {
         // reference input
         let id = RequestId(vec![1]);
         let enr_seq = 1;
-        let message = Payload::Request(Request {
+        let message = Message::Request(Request {
             id,
             body: RequestBody::Ping { enr_seq },
         });
@@ -257,7 +129,7 @@ mod tests {
         // reference input
         let id = RequestId(vec![1]);
         let distances = vec![256];
-        let message = Payload::Request(Request {
+        let message = Message::Request(Request {
             id,
             body: RequestBody::FindNode { distances },
         });
@@ -276,7 +148,7 @@ mod tests {
         let enr_seq = 1;
         let ip: IpAddr = "127.0.0.1".parse().unwrap();
         let port = 5000;
-        let message = Payload::Response(Response {
+        let message = Message::Response(Response {
             id,
             body: ResponseBody::Pong { enr_seq, ip, port },
         });
@@ -297,7 +169,7 @@ mod tests {
         // expected hex output
         let expected_output = hex::decode("04c30101c0").unwrap();
 
-        let message = Payload::Response(Response {
+        let message = Message::Response(Response {
             id,
             body: ResponseBody::Nodes {
                 total,
@@ -317,7 +189,7 @@ mod tests {
         // expected hex output
         let expected_output = hex::decode("04f87b0101f877f875b84028df8ee09f4a62091f1a8b61f18aad2cfe3eadc6f350d527077f9aebc56098883a47f46c6b49bbb91954238f9e14933277b2bd1e0c45b1771b94cd078c32dacb0182696482763489736563703235366b31a103ca634cae0d49acb401d8a4c6b6fe8c55b70d115bf400769cc1400f3258cd3138").unwrap();
 
-        let message = Payload::Response(Response {
+        let message = Message::Response(Response {
             id,
             body: ResponseBody::Nodes {
                 total,
@@ -340,7 +212,7 @@ mod tests {
         // expected hex output
         let expected_output = hex::decode("04f8f20101f8eef875b8401ce2991c64993d7c84c29a00bdc871917551c7d330fca2dd0d69c706596dc655448f030b98a77d4001fd46ae0112ce26d613c5a6a02a81a6223cd0c4edaa53280182696482763489736563703235366b31a103ca634cae0d49acb401d8a4c6b6fe8c55b70d115bf400769cc1400f3258cd3138f875b840d7f1c39e376297f81d7297758c64cb37dcc5c3beea9f57f7ce9695d7d5a67553417d719539d6ae4b445946de4d99e680eb8063f29485b555d45b7df16a1850130182696482763489736563703235366b31a1030e2cb74241c0c4fc8e8166f1a79a05d5b0dd95813a74b094529f317d5c39d235").unwrap();
 
-        let message = Payload::Response(Response {
+        let message = Message::Response(Response {
             id,
             body: ResponseBody::Nodes {
                 total,
@@ -358,10 +230,10 @@ mod tests {
         let expected_enr1 = "enr:-HW4QBzimRxkmT18hMKaAL3IcZF1UcfTMPyi3Q1pxwZZbcZVRI8DC5infUAB_UauARLOJtYTxaagKoGmIjzQxO2qUygBgmlkgnY0iXNlY3AyNTZrMaEDymNMrg1JrLQB2KTGtv6MVbcNEVv0AHacwUAPMljNMTg".parse::<Enr<CombinedKey>>().unwrap();
         let expected_enr2 = "enr:-HW4QNfxw543Ypf4HXKXdYxkyzfcxcO-6p9X986WldfVpnVTQX1xlTnWrktEWUbeTZnmgOuAY_KUhbVV1Ft98WoYUBMBgmlkgnY0iXNlY3AyNTZrMaEDDiy3QkHAxPyOgWbxp5oF1bDdlYE6dLCUUp8xfVw50jU".parse::<Enr<CombinedKey>>().unwrap();
 
-        let decoded = Payload::decode(&input).unwrap();
+        let decoded = Message::decode(&input).unwrap();
 
         match decoded {
-            Payload::Response(response) => match response.body {
+            Message::Response(response) => match response.body {
                 ResponseBody::Nodes { total, nodes } => {
                     assert_eq!(total, 1);
                     assert_eq!(nodes[0], expected_enr1);
@@ -376,13 +248,13 @@ mod tests {
     #[test]
     fn encode_decode_ping_request() {
         let id = RequestId(vec![1]);
-        let request = Payload::Request(Request {
+        let request = Message::Request(Request {
             id,
             body: RequestBody::Ping { enr_seq: 15 },
         });
 
         let encoded = request.clone().encode();
-        let decoded = Payload::decode(&encoded).unwrap();
+        let decoded = Message::decode(&encoded).unwrap();
 
         assert_eq!(request, decoded);
     }
@@ -390,7 +262,7 @@ mod tests {
     #[test]
     fn encode_decode_ping_response() {
         let id = RequestId(vec![1]);
-        let request = Payload::Response(Response {
+        let request = Message::Response(Response {
             id,
             body: ResponseBody::Pong {
                 enr_seq: 15,
@@ -400,7 +272,7 @@ mod tests {
         });
 
         let encoded = request.clone().encode();
-        let decoded = Payload::decode(&encoded).unwrap();
+        let decoded = Message::decode(&encoded).unwrap();
 
         assert_eq!(request, decoded);
     }
@@ -408,7 +280,7 @@ mod tests {
     #[test]
     fn encode_decode_find_node_request() {
         let id = RequestId(vec![1]);
-        let request = Payload::Request(Request {
+        let request = Message::Request(Request {
             id,
             body: RequestBody::FindNode {
                 distances: vec![12],
@@ -416,7 +288,7 @@ mod tests {
         });
 
         let encoded = request.clone().encode();
-        let decoded = Payload::decode(&encoded).unwrap();
+        let decoded = Message::decode(&encoded).unwrap();
 
         assert_eq!(request, decoded);
     }
@@ -441,7 +313,7 @@ mod tests {
 
         let enr_list = vec![enr1, enr2, enr3];
         let id = RequestId(vec![1]);
-        let request = Payload::Response(Response {
+        let request = Message::Response(Response {
             id,
             body: ResponseBody::Nodes {
                 total: 1,
@@ -450,7 +322,7 @@ mod tests {
         });
 
         let encoded = request.clone().encode();
-        let decoded = Payload::decode(&encoded).unwrap();
+        let decoded = Message::decode(&encoded).unwrap();
 
         assert_eq!(request, decoded);
     }
@@ -458,7 +330,7 @@ mod tests {
     #[test]
     fn encode_decode_talk_request() {
         let id = RequestId(vec![1]);
-        let request = Payload::Request(Request {
+        let request = Message::Request(Request {
             id,
             body: RequestBody::TalkReq {
                 protocol: vec![17u8; 32],
@@ -467,7 +339,7 @@ mod tests {
         });
 
         let encoded = request.clone().encode();
-        let decoded = Payload::decode(&encoded).unwrap();
+        let decoded = Message::decode(&encoded).unwrap();
 
         assert_eq!(request, decoded);
     }
