+++ src/rpc/response.rs
@@ -1,13 +1,9 @@
-use super::{RequestBody, RequestId, NODES_MSG_TYPE, PONG_MSG_TYPE, TALKRESP_MSG_TYPE};
+use super::{Payload, RequestBody, RequestId, NODES_MSG_TYPE, PONG_MSG_TYPE, TALKRESP_MSG_TYPE};
 use crate::Enr;
 use parse_display_derive::Display;
 use rlp::{DecoderError, Rlp, RlpStream};
 use std::net::{IpAddr, Ipv6Addr};
 use tracing::debug;
-
-mod response_body;
-
-pub use response_body::ResponseBody;
 
 /// A response sent in response to a [`super::Request`]
 #[derive(Debug, Clone, PartialEq, Eq, Display)]
@@ -19,8 +15,9 @@ pub struct Response {
     pub body: ResponseBody,
 }
 
-impl Response {
-    pub fn msg_type(&self) -> u8 {
+impl Payload for Response {
+    /// Matches a response type to its message type id.
+    fn msg_type(&self) -> u8 {
         match &self.body {
             ResponseBody::Pong { .. } => PONG_MSG_TYPE,
             ResponseBody::Nodes { .. } => NODES_MSG_TYPE,
@@ -28,19 +25,8 @@ impl Response {
         }
     }
 
-    /// Determines if the response is a valid response to the given request.
-    pub fn match_request(&self, req: &RequestBody) -> bool {
-        match self.body {
-            ResponseBody::Pong { .. } => matches!(req, RequestBody::Ping { .. }),
-            ResponseBody::Nodes { .. } => {
-                matches!(req, RequestBody::FindNode { .. })
-            }
-            ResponseBody::TalkResp { .. } => matches!(req, RequestBody::TalkReq { .. }),
-        }
-    }
-
     /// Encodes a response message to RLP-encoded bytes.
-    pub fn encode(self) -> Vec<u8> {
+    fn encode(self) -> Vec<u8> {
         let mut buf = Vec::with_capacity(10);
         let msg_type = self.msg_type();
         buf.push(msg_type);
@@ -88,7 +74,7 @@ impl Response {
     }
 
     /// Decodes RLP-encoded bytes into a response message.
-    pub fn decode(msg_type: u8, rlp: &Rlp<'_>) -> Result<Self, DecoderError> {
+    fn decode(msg_type: u8, rlp: &Rlp<'_>) -> Result<Self, DecoderError> {
         let list_len = rlp.item_count()?;
         let id = RequestId::decode(rlp.val_at::<Vec<u8>>(0)?)?;
         let response = match msg_type {
@@ -176,8 +162,73 @@ impl Response {
                     body: ResponseBody::TalkResp { response },
                 }
             }
-            _ => return Err(DecoderError::Custom("Unknown RPC response message type")),
+            _ => unreachable!("Implementation does not adhere to wire protocol"),
         };
         Ok(response)
+    }
+}
+
+impl Response {
+    /// Determines if the response is a valid response to the given request.
+    pub fn match_request(&self, req: &RequestBody) -> bool {
+        match self.body {
+            ResponseBody::Pong { .. } => matches!(req, RequestBody::Ping { .. }),
+            ResponseBody::Nodes { .. } => {
+                matches!(req, RequestBody::FindNode { .. })
+            }
+            ResponseBody::TalkResp { .. } => matches!(req, RequestBody::TalkReq { .. }),
+        }
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum ResponseBody {
+    /// A PONG response.
+    Pong {
+        /// The current ENR sequence number of the responder.
+        enr_seq: u64,
+        /// Our external IP address as observed by the responder.
+        ip: IpAddr,
+        /// Our external UDP port as observed by the responder.
+        port: u16,
+    },
+    /// A NODES response.
+    Nodes {
+        /// The total number of responses that make up this response.
+        total: u64,
+        /// A list of ENR's returned by the responder.
+        nodes: Vec<Enr>,
+    },
+    /// The TALK response.
+    TalkResp {
+        /// The response for the talk.
+        response: Vec<u8>,
+    },
+}
+
+impl std::fmt::Display for ResponseBody {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            ResponseBody::Pong { enr_seq, ip, port } => {
+                write!(f, "PONG: Enr-seq: {enr_seq}, Ip: {ip:?},  Port: {port}")
+            }
+            ResponseBody::Nodes { total, nodes } => {
+                write!(f, "NODES: total: {total}, Nodes: [")?;
+                let mut first = true;
+                for id in nodes {
+                    if !first {
+                        write!(f, ", {id}")?;
+                    } else {
+                        write!(f, "{id}")?;
+                    }
+                    first = false;
+                }
+
+                write!(f, "]")
+            }
+            ResponseBody::TalkResp { response } => {
+                write!(f, "Response: Response {}", hex::encode(response))
+            }
+        }
     }
 }
